requirements.txt
django<2
Pillow
django-disqus
cryptography
django-hitcount
django-formset-js
